feat: print "registered" as status when using SCA
The message printed for the "overall status" comes directly from the
compliance status of the system. Compliance is a concept strictly
specific to entitlement mode, and in case of SCA there is nothing to be
said, and thus the "unknown" status gets used. While this is technically
correct, in practice it tends to confuse users a lot, as it does not
clearly give the important detail about the system properly registered.

Hence, in case the content access mode is really SCA (after checking the
certificates, potentially refreshing caches and status), use the simple
status string "registered" for a system in SCA mode. This, together with
the message already printed out about the fact that the host should get
access to content, should improve the feedback that users get when
reading the output of "status" on a system registered to an SCA
organization.

Adapt the existing test to the new messaging.